Swap from enum based sampling to class based sampling due to lack of new instancing and cloning...
Added startup pack creation and validation...

+++ VSPE/src/main/java/org/vicky/vspe/utilities/Math/Vector3.java
@@ -1,25 +1,25 @@
 package org.vicky.vspe.utilities.Math;
 
 public class Vector3 {
-   public int x;
-   public int y;
-   public int z;
+    public int x;
+    public int y;
+    public int z;
 
-   public Vector3(int x, int y, int z) {
-      this.x = x;
-      this.y = y;
-      this.z = z;
-   }
+    public Vector3(int x, int y, int z) {
+        this.x = x;
+        this.y = y;
+        this.z = z;
+    }
 
-   public Vector3 add(Vector3 other) {
-      return new Vector3(this.x + other.x, this.y + other.y, this.z + other.z);
-   }
+    public Vector3 add(Vector3 other) {
+        return new Vector3(this.x + other.x, this.y + other.y, this.z + other.z);
+    }
 
-   public Vector3 subtract(Vector3 other) {
-      return new Vector3(this.x - other.x, this.y - other.y, this.z - other.z);
-   }
+    public Vector3 subtract(Vector3 other) {
+        return new Vector3(this.x - other.x, this.y - other.y, this.z - other.z);
+    }
 
-   public double distanceTo(Vector3 other) {
-      return Math.sqrt(Math.pow((double)(this.x - other.x), 2.0) + Math.pow((double)(this.y - other.y), 2.0) + Math.pow((double)(this.z - other.z), 2.0));
-   }
+    public double distanceTo(Vector3 other) {
+        return Math.sqrt(Math.pow(this.x - other.x, 2.0) + Math.pow(this.y - other.y, 2.0) + Math.pow(this.z - other.z, 2.0));
+    }
 }
